requirements.txt
# Core pipeline (matches test.ipynb)
youtube-transcript-api>=0.6.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0

# App
streamlit>=1.39.0
python-dotenv>=1.0.0
